修改, 现在EventDrivenViewModel只会注册来自ViewModel层的监听者, 这样做可以保证ViewModel的改动都可以放在UI线程中来做
